Split chatapp projections and helpers
